# 8341697: C2: Register allocation inefficiency in tight loop