--- a/src/hotspot/share/opto/c2_globals.hpp
+++ b/src/hotspot/share/opto/c2_globals.hpp
@@ -781,6 +781,10 @@
           "Run DuplicateBackedge whenever possible ignoring benefit"        \
           "analysis")                                                       \
                                                                             \
+  product(bool, LoopAwaredSpilling, true, DIAGNOSTIC,                       \
+          "Eagerly spill or reload live ranges at loop entry if it seems"   \
+          "beneficial")                                                     \
+                                                                            \
   product(bool, VerifyReceiverTypes, trueInDebug, DIAGNOSTIC,               \
           "Verify receiver types at runtime")                               \
                                                                             \