… test/micro/org/openjdk/bench/vm/compiler/LoopCounterBench.java