@merykitty 2024-10-11
The benchmark result:

    Benchmark                          Mode  Cnt    Score     Error  Units
    LoopCounterBench.field_ret         avgt    3  417.865 ±   2.914  ns/op
    LoopCounterBench.localVar_ret      avgt    3  332.657 ± 109.310  ns/op

The inner loop is free of spills because it has been hoisted to the loop entry:

                  │  0x00007fdf9821b546:   mov    r9d,DWORD PTR [r11+0xc]      ;*getfield increment {reexecute=0 rethrow=0 return_oop=0}
                  │                                                            ; - org.openjdk.bench.vm.compiler.LoopCounterBench::localVar_ret@1 (line 56)
                  │                                                            ; - org.openjdk.bench.vm.compiler.jmh_generated.LoopCounterBench_localVar_ret_jmhTest::localVar_ret_avgt_jmhStub@17 (line 190)
       0.03%      │  0x00007fdf9821b54a:   mov    esi,DWORD PTR [r12+r8*8+0xc] ; implicit exception: dispatches to 0x00007fdf9821b6f4
                  │                                                            ;*lastore {reexecute=0 rethrow=0 return_oop=0}
                  │                                                            ; - org.openjdk.bench.vm.compiler.LoopCounterBench::localVar_ret@27 (line 58)
                  │                                                            ; - org.openjdk.bench.vm.compiler.jmh_generated.LoopCounterBench_localVar_ret_jmhTest::localVar_ret_avgt_jmhStub@17 (line 190)
                  │  0x00007fdf9821b54f:   lea    rax,[r12+r14*8]
                  │  0x00007fdf9821b553:   lea    r13,[r12+r8*8]
       0.03%      │  0x00007fdf9821b557:   xor    edi,edi
                  │  0x00007fdf9821b559:   vmovq  xmm0,rbp
                  │  0x00007fdf9821b55e:   xchg   ax,ax                        ;*aload_0 {reexecute=0 rethrow=0 return_oop=0}
                  │                                                            ; - org.openjdk.bench.vm.compiler.LoopCounterBench::localVar_ret@16 (line 58)
                  │                                                            ; - org.openjdk.bench.vm.compiler.jmh_generated.LoopCounterBench_localVar_ret_jmhTest::localVar_ret_avgt_jmhStub@17 (line 190)
                 ↗│  0x00007fdf9821b560:   cmp    edi,r10d
       1.66%    ╭││  0x00007fdf9821b563:   jae    0x00007fdf9821b587
                │││  0x00007fdf9821b565:   mov    rbp,QWORD PTR [rax+rdi*8+0x10];*laload {reexecute=0 rethrow=0 return_oop=0}
                │││                                                            ; - org.openjdk.bench.vm.compiler.LoopCounterBench::localVar_ret@26 (line 58)
                │││                                                            ; - org.openjdk.bench.vm.compiler.jmh_generated.LoopCounterBench_localVar_ret_jmhTest::localVar_ret_avgt_jmhStub@17 (line 190)
       5.43%    │││  0x00007fdf9821b56a:   cmp    edi,esi
       0.17%    │││  0x00007fdf9821b56c:   jae    0x00007fdf9821b5c8
                │││  0x00007fdf9821b56e:   mov    QWORD PTR [r13+rdi*8+0x10],rbp;*goto {reexecute=0 rethrow=0 return_oop=0}
                │││                                                            ; - org.openjdk.bench.vm.compiler.LoopCounterBench::localVar_ret@32 (line 57)
                │││                                                            ; - org.openjdk.bench.vm.compiler.jmh_generated.LoopCounterBench_localVar_ret_jmhTest::localVar_ret_avgt_jmhStub@17 (line 190)
       1.40%    │││  0x00007fdf9821b573:   add    edi,r9d                      ;*iadd {reexecute=0 rethrow=0 return_oop=0}
                │││                                                            ; - org.openjdk.bench.vm.compiler.LoopCounterBench::localVar_ret@30 (line 57)
                │││                                                            ; - org.openjdk.bench.vm.compiler.jmh_generated.LoopCounterBench_localVar_ret_jmhTest::localVar_ret_avgt_jmhStub@17 (line 190)
       3.40%    │││  0x00007fdf9821b576:   mov    rbp,QWORD PTR [r15+0x450]    ; ImmutableOopMap {r11=Oop r8=NarrowOop rcx=Oop rbx=Oop rdx=Oop rax=Oop r13=Oop r14=NarrowOop }
                │││                                                            ;*goto {reexecute=1 rethrow=0 return_oop=0}
                │││                                                            ; - (reexecute) org.openjdk.bench.vm.compiler.LoopCounterBench::localVar_ret@32 (line 57)
                │││                                                            ; - org.openjdk.bench.vm.compiler.jmh_generated.LoopCounterBench_localVar_ret_jmhTest::localVar_ret_avgt_jmhStub@17 (line 190)
       1.80%    │││  0x00007fdf9821b57d:   test   DWORD PTR [rbp+0x0],eax      ;*goto {reexecute=0 rethrow=0 return_oop=0}
                │││                                                            ; - org.openjdk.bench.vm.compiler.LoopCounterBench::localVar_ret@32 (line 57)
                │││                                                            ; - org.openjdk.bench.vm.compiler.jmh_generated.LoopCounterBench_localVar_ret_jmhTest::localVar_ret_avgt_jmhStub@17 (line 190)
                │││                                                            ;   {poll}
      84.42%    │││  0x00007fdf9821b580:   cmp    edi,r10d
       0.30%    │╰│  0x00007fdf9821b583:   jl     0x00007fdf9821b560           ;*if_icmpge {reexecute=0 rethrow=0 return_oop=0}
                │ │                                                            ; - org.openjdk.bench.vm.compiler.LoopCounterBench::localVar_ret@13 (line 57)
                │ │                                                            ; - org.openjdk.bench.vm.compiler.jmh_generated.LoopCounterBench_localVar_ret_jmhTest::localVar_ret_avgt_jmhStub@17 (line 190)