@merykitty 2024-10-11
Thanks to @shipilev for the benchmark, could you verify that this can solve the issue in the original benchmark as I imagine this is a simplified version.