@shipilev 2024-10-13
This was really found by @rschwietzke, maybe he would like to test it :)