@rschwietzke 2024-10-14
Fix confirmed. Performance matches the user expectation when pulling data local. I will look into the runtime difference for the plain loop and systemcopy.

### Old - JDK 23.0.0
```
Benchmark                                   (SIZE)  Mode  Cnt     Score     Error  Units
Example8ArrayCopying.manualCopy1              1000  avgt   10    70.222 ±   3.549  ns/op
Example8ArrayCopying.manualCopy2              1000  avgt   10    70.011 ±   0.880  ns/op
Example8ArrayCopying.manualCopyAntiUnroll1    1000  avgt   10   394.275 ±  20.067  ns/op
Example8ArrayCopying.manualCopyAntiUnroll2    1000  avgt   10   636.158 ± 101.505  ns/op
Example8ArrayCopying.manualCopyAntiUnroll3    1000  avgt   10  1646.330 ±  23.042  ns/op
Example8ArrayCopying.systemCopy               1000  avgt   10    74.845 ±   1.535  ns/op
```

### New - JDK 24-internal (merrykitty/improveregalloc, 12d1a2b21fc62145dac04fecf43f267f539b2aa5)
```
Example8ArrayCopying.manualCopy1              1000  avgt   10   80.155 ±  4.504  ns/op
Example8ArrayCopying.manualCopy2              1000  avgt   10   81.122 ±  3.074  ns/op
Example8ArrayCopying.manualCopyAntiUnroll1    1000  avgt   10  394.094 ±  6.809  ns/op
Example8ArrayCopying.manualCopyAntiUnroll2    1000  avgt   10  626.155 ± 13.055  ns/op
Example8ArrayCopying.manualCopyAntiUnroll3    1000  avgt   10  564.199 ± 23.854  ns/op
Example8ArrayCopying.systemCopy               1000  avgt   10   99.393 ±  0.634  ns/op
```

Source code for reference: https://github.com/Xceptance/jmh-training/blob/1dbcc9c38553b0e8b683c6f70475a25150b66635/src/main/java/org/xc/jmh/Example8ArrayCopying.java